\name{full_experimental}
\alias{full_experimental}
\title{Double a number}
\usage{
full_experimental(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
\code{x * 2}.
}
\description{
\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}

Doubles its input. The interface may still change.
}
\examples{
full_experimental(21)
}
