\name{+.fullclass}
\alias{+.fullclass}
\title{Add two fullclass objects}
\usage{
\method{+}{fullclass}(e1, e2)
}
\arguments{
Both operands are coerced with \code{unclass()} before addition.

\item{e1, e2}{Objects of class \code{fullclass}.}
}
\value{
A \code{fullclass} object.
}
\description{
Adds two \code{fullclass} objects elementwise.
}
